+++ src/main/java/com/sdy/bbbb/data/DataController.java
@@ -35,4 +35,5 @@ public class DataController {
         return dataService.getPopulationChanges();
     }
 
+
 }
