+++ src/main/java/com/sdy/bbbb/data/dataDto/JamDto.java
@@ -3,9 +3,8 @@ package com.sdy.bbbb.data.dataDto;
 public interface JamDto {
 
     Long getWw();
-
     String getArea_Nm();
-
+    String getGu_Nm();
     Long getScore_Num();
 
 }
