+++ src/main/java/com/sdy/bbbb/data/dataDto/JamTop5Dto.java
@@ -7,17 +7,20 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class JamTop5Dto {
 
-//    private Long ww;
+    private Long ww;
 
     private String areaNm;
 
     private Long scoreSum;
 
+    private String guNm;
+
 
     public JamTop5Dto(JamDto jamDto) {
-//        this.ww = jamDto.getWw();
+        this.ww = jamDto.getWw();
         this.areaNm = jamDto.getArea_Nm();
         this.scoreSum = jamDto.getScore_Num();
+        this.guNm = jamDto.getGu_Nm();
     }
 }
