+++ src/main/java/com/sdy/bbbb/data/dataDto/PopulationDto.java
@@ -14,4 +14,6 @@ public interface PopulationDto {
 
     String getArea_Congest_Lvl();
 
+    String getGu_Nm();
+
 }
